[cuda] FAST stops losing on memory: 1,072 KB to 440 KB
FAST led cv::cuda::FastFeatureDetector on speed by 5.8x and LOST on memory,
1,088 KB against 704 KB. That should not happen to a library holding one bit
per pixel, and it was not the representation: the bit plane is 45 KB of that
figure. The other 94% was keypoint capacity, and two of its terms were ours.

THE SCRATCH WAS SIZED FOR AN ARM THAT DOES NOT RUN. fastScratchBytes returned
the reference sort arm's padded working area -- nextPow2(capacity) records --
whichever arm was selected, so production allocated 512 KB to keep reachable a
kernel the ordered arm replaced. The ordered arm needs one uint32_t per block:
380 bytes for a 752x480 frame.

The reason that number did not vary was sound and is kept: a sizing function
whose answer moves when an implementation detail moves turns a correct caller
into a device-side out-of-bounds write with no signal. The two arms' scratch are
functions of DIFFERENT THINGS, though -- the reference arm's of capacity, the
ordered arm's of the frame -- so one number could only ever be their maximum.
The arm is now an explicit parameter, so the answer is a pure function of the
caller's own arguments, and detectFastAsync validates against the arm it is
ABOUT to run rather than the one the caller sized for. The unsafe combination
returns cudaErrorInvalidValue. Proven, not documented: against poisoned 1 MB
guards either side, the refused call disturbs zero bytes in the guards, in the
380-byte window and in the output buffer, and leaves the counter at zero.

Sizing by the scratchBytes actually passed was considered and rejected outright
-- that makes the arm depend on the allocation, so an under-allocating caller
silently gets a different, slower, differently-truncating arm instead of an
error. That is the original hazard inverted.

THE RECORD NARROWS 16 BYTES TO 12. DeviceFastCorner carried a long long score,
which forced 8-byte alignment and four bytes of pure padding. The 64 bits are
inherited rather than needed: the device ports the BIT-PLANE overload, whose
score is fastLongestRun over a 16-bit ring with attainable set [1, 16]. The
host's wide-image overload sums ring differences over an unconstrained source
and reaches about 2^36, so the HOST type genuinely needs 64 bits and is
untouched.

The narrowing is proven exhaustively rather than sampled: all 65,536 ring
patterns x all 16 arc lengths, every value stored narrow and read back through
toHost. Positions do NOT narrow -- cv::cuda reaches 12 bytes with a short2 and
a 32,767-pixel frame cap, and a position is the answer.

  bit plane                          46,080 B      46,080 B
  corner array    32,768 x 16 =     524,288 B  ->  32,768 x 12 = 393,216 B
  scratch    nextPow2(32768) x 16 = 524,288 B  ->  95 blocks x 4 =    380 B
  metered, cudaMemGetInfo, both sides, 256 replicas
                                  1,072.0 KB   ->            440.0 KB

Against OpenCV's 680-704 KB on the same meter that is 1.574x SMALLER where it
was 1.545x larger. Speed did not pay for it: 0.1664x, 6.01x faster, disjoint
7/7, against a standing 0.171x at 6/7.

Three corrections the verification made to its own inputs:

  * The first memory reading was under-resolved. At 64 replicas one meter unit
    is 32 KB per frame on a 430 KB reading, and two harnesses that both cleared
    the eight-unit rule still disagreed (1.615x against 1.714x). At 256
    replicas the unit is 8 KB and both stop moving.
  * OpenCV's FAST working set is CONTENT-DEPENDENT -- it sizes its output mat by
    corners found, so a synthetic picture reads 576 KB where the real edge map
    reads 704. This row cannot be measured on synthetic content.
  * The header claimed arm selection was unchanged because the comparison was
    between the same two numbers. It was not: narrowing the record moved one
    side by a quarter, so 4.5% of (frame, capacity) points now take the
    reference arm, always that direction and never above capacity 128. Both
    arms are bit-exact so no answer changes, but a byte count is not a work
    unit and the header said otherwise.

The profiler records a cost that is free today and will not stay free: a
12-byte record at 4-byte alignment is three scattered stores where 16 bytes at
8-byte alignment packed into wider ones, and L1 global store sectors rise 42%.
Duration, SM throughput and achieved occupancy are unchanged to three figures
because this kernel is occupancy-bound, which is #64's subject.

Also in this pass: goodFeaturesToTrack moves out of the undelivered list and
into the entries that lead on both axes -- 5.3x faster, 5.33x smaller, disjoint
7/7 -- with the caveat that OpenCV's arm swings 3.24-13.61 ms because its
min-distance spacing runs on the CPU. The census entry's memory loss is
recorded as INHERENT rather than as a defect: census expands 8 bits per pixel
into a 32-bit descriptor word, so two transformed images are 2,820 KB before
anything else. The plane layout is offered as the documented alternative at
3,217 KB against packed's 3,877, and its cost is re-measured at 21.7x rather
than the stale 8.55x, which predates the warp-box matcher.

README's FAST paragraph was two rounds stale, still claiming a 16.8x miss and a
frontend 1.22x slower than the host. Rewritten.

verify_cuda.sh: 194,986 Release / 194,933 Debug across 17 suites.

+++ backends/cuda/tests/test_cuda_keypoints.cpp
@@ -175,7 +175,7 @@ BINCV_TEST(CudaKeypointsFromCorners, TheFastCornerSpellingAgreesWithTheHostConve
     for (uint32_t i = 0; i < capacity; ++i) {
         fast[i].x = static_cast<int>(splitmix(seed) % 752u);
         fast[i].y = static_cast<int>(splitmix(seed) % 480u);
-        fast[i].score = static_cast<long long>(splitmix(seed) % 16u);
+        fast[i].score = static_cast<int>(splitmix(seed) % 16u);
         mirror[i].x = fast[i].x;
         mirror[i].y = fast[i].y;
         mirror[i].response = 0.0f;
